Extra documentation for processListings
Mostly to clarify that it is a wrapper to make it easy to do the weekend
listings (two sets on one page, requiring different frame names) and not
a ginormous, unwieldy function.

I was concerned at its size and scope, but have managed to set my mind at
ease with the wrapper rationale. An audit all of the script's functions
showed that they are mostly short and always task-oriented.

So now I feel much better.